calendar_tracker/requirements.txt
flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.0.1
google-generativeai==0.8.4
pydantic==2.10.6
jinja2==3.1.6
werkzeug==3.1.3
click==8.1.8
